requirements.txt
python-decouple
PyMySQL
requests